requirements.txt
absl-py==1.4.0
altair==4.2.2
attrs==23.1.0
backcall==0.2.0
base58==2.1.1
bleach==6.0.0
blinker==1.6.2
boto3==1.28.0
botocore==1.31.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
docutils==0.16
entrypoints==0.4
enum-compat==0.0.3
future==0.18.3
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.57.0
gunicorn==21.2.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.12.2
ipywidgets==8.1.1
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
jupyter-client==7.4.9
jupyter-core==5.5.0
Keras==2.13.1
Markdown==3.4.3
MarkupSafe==2.1.3
mistune==3.0.1
nbconvert==7.11.0
nbformat==5.9.2
notebook==7.0.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==10.0.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1
Pygments==2.16.1
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
streamlit==1.26.0
tensorboard==2.13.0
tensorflow==2.13.1
tensorflow-estimator==2.13.0
termcolor==2.3.0
terminado==0.18.0
testpath==0.6.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.10.1
tzlocal==5.0.1
urllib3==2.0.4
validators==0.21.0
watchdog==3.0.0
wcwidth==0.2.8
webencodings==0.5.1
Werkzeug==2.3.7
widgetsnbextension==4.0.9
wrapt==1.15.0
zipp==3.16.2
